tracking_service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
StrEnum>=0.4.15
